First pass at incorporating all the motor files into build for new Tiva chip. Note that there are 3 separate builds that can be done here: 1. Cygwin, 2. Stellaris (the old controller) and 3. Tiva There are minor changes between the Stellaris and Tiva build and those can be found by searching on the preprocessor: PART_TM4C1233D5PM One task still to do is to modify usb_dev_serial.c to accept the motor control program. The old UART code from the Stellaris build is still there, but should be #ifdef'd out for the Tiva part. Saving the program to flash also still needs implemented and tested.

+++ flashio.cpp
@@ -5,17 +5,19 @@
 
 */
 
-#include "flashio.h"
+#include <stdbool.h>
+#include <stdint.h>
 #include "hw_ints.h"
 #include "hw_memmap.h"
 #include "hw_types.h"
-#include "/DriverLib/src/flash.h"   // Using "flash.h" pulls in file from FreeRTOS
+#include "driverlib/flash.h"   // Using "flash.h" pulls in file from FreeRTOS
+#include "flashio.h"
 
 flashio::flashio() {
 }
 
-void flashio::saveData(unsigned long* pData, unsigned long count) {
-    unsigned long* remainingDataPointer = pData;
+void flashio::saveData(uint32_t* pData, unsigned long count) {
+    uint32_t* remainingDataPointer = pData;
     unsigned long remainingCount = count;
     unsigned int pageCount = 0;
     while (remainingCount > 0) {
